feat(cassandraTest): add methods to create and delete keyspace
Rename FistCassandraTest.go to cassandraTest.go.
Add function to create and drop keyspace.
